+++ src/Execution/Arguments/ArgumentSet.php
@@ -2,13 +2,6 @@
 
 namespace Nuwave\Lighthouse\Execution\Arguments;
 
-use Nuwave\Lighthouse\Execution\ResolveInfo;
-use Nuwave\Lighthouse\Scout\ScoutEnhancer;
-use Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective;
-use Nuwave\Lighthouse\Support\Contracts\FieldBuilderDirective;
-use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
-use Nuwave\Lighthouse\Support\Utils;
-
 class ArgumentSet
 {
     /**
@@ -64,104 +57,6 @@ class ArgumentSet
         }
 
         return null !== $argument->value;
-    }
-
-    /**
-     * Apply ArgBuilderDirectives and scopes to the builder.
-     *
-     * @template TBuilder of \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Relations\Relation|\Laravel\Scout\Builder
-     *
-     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Relations\Relation|\Laravel\Scout\Builder  $builder
-     *
-     * @phpstan-param  TBuilder  $builder
-     *
-     * @param  array<string>  $scopes
-     * @param  array<string, mixed>  $args
-     *
-     * @return \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder|\Illuminate\Database\Eloquent\Relations\Relation|\Laravel\Scout\Builder
-     *
-     * @phpstan-return TBuilder
-     *
-     * @deprecated will be moved to \Nuwave\Lighthouse\Execution\ResolveInfo
-     */
-    public function enhanceBuilder(object $builder, array $scopes, $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo, \Closure $directiveFilter = null): object
-    {
-        $scoutEnhancer = new ScoutEnhancer($this, $builder);
-        if ($scoutEnhancer->canEnhanceBuilder()) {
-            return $scoutEnhancer->enhanceBuilder();
-        }
-
-        self::applyArgBuilderDirectives($this, $builder, $directiveFilter);
-        self::applyFieldBuilderDirectives($builder, $root, $args, $context, $resolveInfo);
-
-        foreach ($scopes as $scope) {
-            $builder->{$scope}($this->toArray());
-        }
-
-        return $builder;
-    }
-
-    /**
-     * Recursively apply the ArgBuilderDirectives onto the builder.
-     *
-     * TODO get rid of the reference passing in here. The issue is that @search makes a new builder instance,
-     * but we must special case that in some way anyhow, as only eq filters can be added on top of search.
-     *
-     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
-     * @param  (\Closure(\Nuwave\Lighthouse\Support\Contracts\ArgBuilderDirective): bool)|null  $directiveFilter
-     */
-    protected static function applyArgBuilderDirectives(self $argumentSet, object &$builder, \Closure $directiveFilter = null): void
-    {
-        $unboxBenSampoEnumEnumInstances = config('lighthouse.unbox_bensampo_enum_enum_instances');
-
-        foreach ($argumentSet->arguments as $argument) {
-            $value = $argument->toPlain();
-
-            // TODO remove in v6, Laravel automagically calls the Enum's __toString() method
-            // Unbox Enum values to ensure their underlying value is used for queries
-            if ($unboxBenSampoEnumEnumInstances && is_a($value, '\BenSampo\Enum\Enum')) {
-                $value = $value->value;
-            }
-
-            $filteredDirectives = $argument
-                ->directives
-                ->filter(Utils::instanceofMatcher(ArgBuilderDirective::class));
-
-            if (null !== $directiveFilter) {
-                $filteredDirectives = $filteredDirectives->filter($directiveFilter);
-            }
-
-            $filteredDirectives->each(static function (ArgBuilderDirective $argBuilderDirective) use (&$builder, $value): void {
-                $builder = $argBuilderDirective->handleBuilder($builder, $value);
-            });
-
-            Utils::applyEach(
-                static function ($value) use (&$builder, $directiveFilter) {
-                    if ($value instanceof self) {
-                        self::applyArgBuilderDirectives($value, $builder, $directiveFilter);
-                    }
-                },
-                $argument->value
-            );
-        }
-    }
-
-    /**
-     * Apply the FieldBuilderDirectives onto the builder.
-     *
-     * TODO get rid of the reference passing in here. The issue is that @search makes a new builder instance,
-     * but we must special case that in some way anyhow, as only eq filters can be added on top of search.
-     *
-     * @param  \Illuminate\Database\Query\Builder|\Illuminate\Database\Eloquent\Builder  $builder
-     * @param  array<string, mixed>  $args
-     */
-    protected static function applyFieldBuilderDirectives(object &$builder, $root, array $args, GraphQLContext $context, ResolveInfo $resolveInfo): void
-    {
-        $resolveInfo->argumentSet->directives
-            ->filter(Utils::instanceofMatcher(FieldBuilderDirective::class))
-            ->each(static function (FieldBuilderDirective $fieldBuilderDirective) use (&$builder, $root, $args, $context, $resolveInfo): void {
-                $builder = $fieldBuilderDirective->handleFieldBuilder($builder, $root, $args, $context, $resolveInfo);
-            });
     }
 
     /**