+++ CHANGELOG.md
@@ -12,6 +12,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Changed
 
 - Pass resolver arguments to `FieldBuilderDirective::handleFieldBuilder()` https://github.com/nuwave/lighthouse/pull/2234
+- Expected resolver arguments in `ResolveInfo::enhanceBuilder()`
 
 ### Fixed
 
@@ -32,6 +33,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Remove support for Laravel 5, 6, 7
 - Remove `Serializable` implementation
 - Remove trait `ClearsSchemaCache`
+- Remove config option `lighthouse.unbox_bensampo_enum_enum_instances`
+- Remove `ArgumentSet::enhanceBuilder()`, use `ResolveInfo::enhanceBuilder()`
 
 ## v5.70.3